Support `kid` to locate the key in a JSON Web Key Set (JWKS)
In `encode/{3,4}`, it's possible to annotate the key with it key ID using a
tuple:

    jwt:encode(Alg, ClaimsSet, {<<"MyKeyId">>, Key}).

The `kid` attribute will be added to the token header.

In `decode/{2,3}`, it's possible to pass an encoded (JSON as a binary)
or decoded (JSON as a map) JWKS:

    jwt:decode(Token, JWKS).

`jwt` will call `jwk` to decode (if necessary) and select the
appropriate key in the JWKS during the `get_key` step, as the `kid` was
obtained previously during the `decode_jwt` step.

As part of that, new type specs were defined:
* `key()`
* `key_id()`

Care has been taken to not break the API in the process.